telegram-desktop: bump
Wide-Baloons patch is kinda "broken" now:
upstream changed the code so heavy that I failed to get the proper hack
to fix odd behaviour when you click on wide-screen checkbox (baloons
change their width, but not height, so text either overlaps the baloon
(if previous state was "widescreen"), or baloon have much bigger height
than it needs, if previous state was "non-widescreen").

As a workaround you can either resize window, or hide&show it (at least,
clicking on the tray icon worked for me).